visit & booking

+++ FitHubBackendAPI/DTOs/GymBranchDTOs/GetGymBranchByIdDTO.cs
@@ -12,7 +12,7 @@ namespace FitHubBackendAPI.DTOs.GymBranchDTOs
         public string? Address { get; set; }
         public string? City { get; set; }
         // price in credits for a single visit without a plan
-        public int VisitCreditsCost { get; set; }
+        public decimal VisitCreditsCost { get; set; }
         public string? Description { get; set; }
 
         public TimeSpan? OpenTime { get; set; }
