+++ FitHubBackendAPI/DTOs/Wallet/TransactionHistoryDto.cs
@@ -7,7 +7,7 @@
         public string Description { get; set; } = string.Empty; // "Pro Plan", "Class Booking"
 
         public decimal AmountPaid { get; set; } // الفلوس اللي دفعها ($49.99)
-        public int Credits { get; set; }        // النقط اللي زادت أو نقصت (+50 / -5)
+        public decimal Credits { get; set; }        // النقط اللي زادت أو نقصت (+50 / -5)
 
         public string Type { get; set; } = string.Empty; // نوع العملية
         public bool IsPositive { get; set; }    // عشان الفرونت يلونها (True=أخضر, False=أحمر)
