+++ FitHubBackendAPI/Entities/Models/UserWallet.cs
@@ -4,7 +4,7 @@
     {
         public int? UserId { get; set; }
 
-        public int? Balance { get; set; }= 0;
+        public decimal? Balance { get; set; }= 0;
         public DateTime? LastUpdated { get; set; }
 
         public User? User { get; set; }
